moved main function so it dows not use the app
