+++ kniteditor/__main__.py
@@ -6,5 +6,5 @@ if __name__ == "__main__":
     import sys
     import os
     sys.path.append(os.path.join(os.path.dirname(__file__), ".."))
-    from kniteditor.EditorWindow import main
+    from kniteditor import main
     main()
